refactor: extract camera-ready preflight orchestration
